fix(jwt): Moving library. Setting credentials in config file
