+++ src/lib/wlogger.js
@@ -44,13 +44,11 @@ var wlogger = winston.createLogger({
 })
 
 // This controls the logs to CONSOLE
-/*
 wlogger.add(
   new winston.transports.Console({
     format: winston.format.simple(),
-    level: "info"
+    level: 'info'
   })
 )
-*/
 
 module.exports = wlogger
